textureデータをゲット
